# openapi-main.yaml
openapi: 3.0.0
servers: 
  - url: 'http://localhost:3000' 
info:
  title: OpenAPI Specification
  version: "1"
...
paths:
...
  /accessories:
    post:
      tags:
        - Accessories
      operationId: createAccessory
      summary: "Create a new accessory"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                magic_defence:
                  type: integer
      responses:
        '201':
          description: Accessory created
        '400':
          description: Invalid input
    get:
      tags:
        - Accessories
      operationId: listAccessories
      summary: "List all accessories"
      parameters:
        - name: "sortBy"
          in: "query"
          description: "Sort by"
          required: false
          schema:
            type: "string"
        - name: "sortDirection"
          in: "query"
          description: "Sort direction"
          required: false
          schema:
            type: "string"
            enum: ["asc", "desc"]
      responses:
        '200':
          description: A list of accessories
  /accessories/{id}:
    get:
      tags:
        - Accessories
      operationId: getAccessoryById
      summary: "Get accessory by id"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of accessory to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                  name:
                    type: string
                  magic_defence:
                    type: integer
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Accessory not found"
    put:
      tags:
        - Accessories
      operationId: updateAccessory
      summary: "Update an existing accessory"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of accessory to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                magic_defence:
                  type: integer
      responses:
        '200':
          description: "successful operation"
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Accessory not found"
    delete:
      tags:
        - Accessories
      operationId: deleteAccessory
      summary: "Delete an accessory"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of accessory to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Accessory not found"
  /accessories/stats:
    post:
      tags:
        - Accessories stats
      operationId: createAccessoryStats
      summary: "Create a new accessory stats"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accessory_id:
                  type: integer
                stat_id:
                  type: integer
                value:
                  type: integer
      responses:
        '201':
          description: Accessory stats created
        '400':
          description: Invalid input
    get:
      tags:
        - Accessories stats
      operationId: listAccessoryStats
      summary: "List all accessory stats"
      parameters:
        - name: "sortBy"
          in: "query"
          description: "Sort by"
          required: false
          schema:
            type: "string"
        - name: "sortDirection"
          in: "query"
          description: "Sort direction"
          required: false
          schema:
            type: "string"
            enum: ["asc", "desc"]
      responses:
        '200':
          description: A list of accessory stats
  /accessories/{id}/stats:
    get:
      tags:
        - Accessories stats
      operationId: getAccessoryStats
      summary: "Get accessory stats"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of accessory to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                  statId:
                    type: integer
                  value:
                    type: integer
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Accessory not found"
    post:
      tags:
        - Accessories
      summary: "Create a new accessory stats"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of accessory to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                statId:
                  type: integer
                value:
                  type: integer          
      responses:
        '201':
          description: Accessory stats created
        '400':
          description: Invalid input
  /accessories/{id}/stats/{stat_id}:
    get:
      tags:
        - Accessories stats
      operationId: getAccessoryStatsById
      summary: "Get accessory stats by id"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of accessory to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "stat_id"
          in: "path"
          description: "ID of accessory stats to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                  statId:
                    type: integer
                  value:
                    type: integer
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Accessory stats not found"
    put:
      tags:
        - Accessories stats
      operationId: updateAccessoryStats
      summary: "Update an existing accessory stats"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of accessory to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "stat_id"
          in: "path"
          description: "ID of accessory stats to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                statId:
                  type: integer
                value:
                  type: integer
      responses:
        '200':
          description: "successful operation"
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Accessory stats not found"
    delete:
      tags:
        - Accessories stats
      operationId: deleteAccessoryStats
      summary: "Delete a accessory stats"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of accessory to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "stat_id"
          in: "path"
          description: "ID of accessory stats to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Accessory stats not found"
  /armors:
    post:
      tags:
        - Armors
      operationId: createArmor
      summary: "Create a new armor"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                melee_defence:
                  type: integer
                ranged_defence:
                  type: integer
      responses:
        '201':
          description: Armor created
        '400':
          description: Invalid input
    get:
      tags:
        - Armors
      operationId: listArmors
      summary: "List all armors"
      parameters:
        - name: "sortBy"
          in: "query"
          description: "Sort by"
          required: false
          schema:
            type: "string"
        - name: "sortDirection"
          in: "query"
          description: "Sort direction"
          required: false
          schema:
            type: "string"
            enum: ["asc", "desc"]
      responses:
        '200':
          description: A list of armors
  /armors/{id}:
    get:
      tags:
        - Armors
      operationId: getArmorById
      summary: "Get armor by id"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of armor to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                  name:
                    type: string
                  melee_defence:
                    type: integer
                  ranged_defence:
                    type: integer
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Armor not found"
    put:
      tags:
        - Armors
      operationId: updateArmor
      summary: "Update an existing armor"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of armor to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                melee_defence:
                  type: integer
                ranged_defence:
                  type: integer
      responses:
        '200':
          description: "successful operation"
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Armor not found"
    delete:
      tags:
        - Armors
      operationId: deleteArmor
      summary: "Delete an armor"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of armor to return"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        '200':
          description: "successful operation"
        '400':
          description: "Invalid ID supplied"
        '404':
          description: "Armor not found"
...